fix: moved job application route
I delteted the job_application.py file and move the job application
route logic ot the job.py file in the route folder.
